+++ CHANGES.md
@@ -38,6 +38,8 @@
 - Fix LWC Debug mode force page refresh before User update [issue 718](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/718) (contribution by [Paul Kalinin](https://github.com/Paul-Kalynyn))
 - Fix `event-monitor` Error on Edge [issue 716](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/716) (contribution by [Paul Kalinin](https://github.com/Paul-Kalynyn))
 - Handle large IN clause queries in `data export` by automatically chunking them into smaller queries [issue 453](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/453) (contribution by [Brahim LAISSAOUI](https://github.com/laissaouibrahim))
+- Filter result by column in `data export`[issue 685](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/685) (contribution by [Brahim LAISSAOUI](https://github.com/laissaouibrahim))
+
 
 ## Version 1.25